Response processing refactor
Move response processing to a separate class
